Upgrades/BitField refactor Upgrades have been refactored to include all the known upgrade types, including those that aren't available from anomalies. Because there are two different index values used depending on whether you're dealing with anomalies or activations, the Upgrades enum now supports both, and the ordinal() value is now meaningless. Because of the huge bit field for updating Upgrade status on a player ship makes creating an enum for it rather onerous, BitField, PacketReader, PacketWriter have been refactored to no longer require an array of enum values (though the methods that support them in PacketReader and PacketWriter still exist).

+++ src/main/java/com/walkertribe/ian/world/ArtemisNpc.java
@@ -37,7 +37,7 @@ public class ArtemisNpc extends BaseArtemisShip {
     public static final byte SCAN_LEVEL_BASIC = 1;
     public static final byte SCAN_LEVEL_FULL  = 2;
 
-    private byte mScanLevel = -1;
+    private int mScanLevel = -1;
     private int mSpecial = -1, mSpecialState = -1;
     private BoolState mEnemy = BoolState.UNKNOWN;
     private BoolState mSurrendered = BoolState.UNKNOWN;
@@ -198,11 +198,11 @@ public class ArtemisNpc extends BaseArtemisShip {
      * The scan level for this ship.
      * Unspecified: -1
      */
-    public byte getScanLevel() {
+    public int getScanLevel() {
         return mScanLevel;
     }
 
-    public void setScanLevel(byte scanLevel) {
+    public void setScanLevel(int scanLevel) {
         mScanLevel = scanLevel;
     }
